Move dataloader logic out of the driver.
