+++ packages/orm/src/collections/OneToManyCollection.ts
@@ -11,9 +11,9 @@ import {
 import { remove } from "../utils";
 import { AbstractRelationImpl } from "./AbstractRelationImpl";
 import { ManyToOneReference } from "./ManyToOneReference";
+import { oneToManyDataLoader } from "../dataloaders/oneToManyDataLoader";
 
-export class OneToManyCollection<T extends Entity, U extends Entity>
-  extends AbstractRelationImpl<U[]>
+export class OneToManyCollection<T extends Entity, U extends Entity> extends AbstractRelationImpl<U[]>
   implements Collection<T, U> {
   private loaded: U[] | undefined;
   // We don't need to track removedBeforeLoaded, because if a child is removed in our unloaded state,
@@ -43,7 +43,7 @@ export class OneToManyCollection<T extends Entity, U extends Entity>
         this.loaded = [];
       } else {
         const em = getEm(this.entity);
-        this.loaded = await em.driver.loadOneToMany(em, this);
+        this.loaded = await oneToManyDataLoader(em.loadLoaders, this).load(this.entity.id);
       }
       this.maybeAppendAddedBeforeLoaded();
     }
@@ -158,7 +158,7 @@ export class OneToManyCollection<T extends Entity, U extends Entity>
     // TODO We should remember what load hints have been applied to this collection and re-apply them.
     if (this.loaded !== undefined && this.entity.id !== undefined) {
       const em = getEm(this.entity);
-      this.loaded = await em.driver.loadOneToMany(em, this);
+      this.loaded = await oneToManyDataLoader(em.loadLoaders, this).load(this.entity.id);
     }
   }
