+++ packages/orm/src/collections/OneToOneReference.ts
@@ -2,6 +2,7 @@ import { deTagIds, ensureNotDeleted, fail, getEm, IdOf, Reference, setField } fr
 import { Entity, EntityMetadata, getMetadata } from "../EntityManager";
 import { AbstractRelationImpl } from "./AbstractRelationImpl";
 import { ManyToOneReference } from "./ManyToOneReference";
+import { oneToOneDataLoader } from "../dataloaders/oneToOneDataLoader";
 
 /**
  * Represents the "many" side of a one-to-one relationship.
@@ -20,8 +21,7 @@ import { ManyToOneReference } from "./ManyToOneReference";
  * Currently we enforce this with a runtime check, which is not great, but the trade-off of implementing
  * `Reference` seemed worth the downside of a un-type-safe `.id` property.
  */
-export class OneToOneReference<T extends Entity, U extends Entity>
-  extends AbstractRelationImpl<U>
+export class OneToOneReference<T extends Entity, U extends Entity> extends AbstractRelationImpl<U>
   implements Reference<T, U, undefined> {
   private loaded: U | undefined;
   private isLoaded: boolean = false;
@@ -72,7 +72,7 @@ export class OneToOneReference<T extends Entity, U extends Entity>
     if (!this.isLoaded) {
       if (!this.entity.isNewEntity) {
         const em = getEm(this.entity);
-        this.loaded = await em.driver.loadOneToOne(em, this);
+        this.loaded = await oneToOneDataLoader(em.loadLoaders, this).load(this.entity.idOrFail);
       }
       this.isLoaded = true;
     }
