+++ packages/orm/src/drivers/driver.ts
@@ -1,34 +1,37 @@
 import Knex from "knex";
-import { ManyToManyCollection } from "../collections/ManyToManyCollection";
+import { JoinRow, ManyToManyCollection } from "../collections/ManyToManyCollection";
 import { OneToManyCollection } from "../collections/OneToManyCollection";
 import { OneToOneReference } from "../collections/OneToOneReference";
 import { Entity, EntityConstructor, EntityManager, EntityMetadata, FilterOf, OrderOf } from "../EntityManager";
 import { JoinRowTodo, Todo } from "./EntityPersister";
+import { FilterAndSettings } from "../QueryBuilder";
 
 export interface Driver {
-  load<T extends Entity>(em: EntityManager, meta: EntityMetadata<T>, id: string): Promise<T | undefined>;
+  load<T extends Entity>(em: EntityManager, meta: EntityMetadata<T>, ids: readonly string[]): Promise<unknown[]>;
 
   loadManyToMany<T extends Entity, U extends Entity>(
     em: EntityManager,
     collection: ManyToManyCollection<T, U>,
-  ): Promise<U[]>;
+    keys: readonly string[],
+  ): Promise<JoinRow[]>;
 
   loadOneToMany<T extends Entity, U extends Entity>(
     em: EntityManager,
     collection: OneToManyCollection<T, U>,
-  ): Promise<U[]>;
+    ids: readonly string[],
+  ): Promise<unknown[]>;
 
   loadOneToOne<T extends Entity, U extends Entity>(
     em: EntityManager,
     reference: OneToOneReference<T, U>,
-  ): Promise<U | undefined>;
+    ids: readonly string[],
+  ): Promise<unknown[]>;
 
   find<T extends Entity>(
     em: EntityManager,
     type: EntityConstructor<T>,
-    where: FilterOf<T>,
-    options?: { orderBy?: OrderOf<T>; limit?: number; offset?: number },
-  ): Promise<unknown[]>;
+    queries: readonly FilterAndSettings<T>[],
+  ): Promise<unknown[][]>;
 
   transaction<T>(
     em: EntityManager,
@@ -39,6 +42,4 @@ export interface Driver {
   flushEntities(todos: Record<string, Todo>): Promise<void>;
 
   flushJoinTables(joinRows: Record<string, JoinRowTodo>): Promise<void>;
-
-  resetDataLoaderCache(): void;
 }
